[EG] Cambiadas ocurrencias donde se usaba Double por Integer

+++ src/AlmacenadorResultados/AlmacenadorResultados.java
@@ -5,17 +5,17 @@ import java.util.Map;
 
 public class AlmacenadorResultados {
 
-    private Map<Integer,Double> resultados;
+    private Map<Integer,Integer> resultados;
 
     public AlmacenadorResultados() {
         this.resultados = new HashMap<>();
     }
 
-    public void agregarResultado(Integer valor, Double distancia) {
+    public void agregarResultado(Integer valor, Integer distancia) {
         this.resultados.put(valor,distancia);
     }
 
-    synchronized public Map<Integer,Double> getResultados() {
+    synchronized public Map<Integer,Integer> getResultados() {
         while(hayWorkersLaburando) {
             try {wait();} catch (Exception e) { return resultados; }
         }
